+++ packages/typespec-client-generator-core/src/package.ts
@@ -59,6 +59,7 @@ import {
 import { createDiagnostic } from "./lib.js";
 import {
   getClientNamespaceString,
+  getCrossLanguageDefinitionId,
   getCrossLanguagePackageId,
   getDefaultApiVersion,
   getEffectivePayloadType,
@@ -286,6 +287,7 @@ function getSdkBasicServiceMethod<
     getResponseMapping: function getResponseMapping(): string | undefined {
       return undefined; // currently we only return a value for paging or lro
     },
+    crossLanguageDefintionId: getCrossLanguageDefinitionId({ ...operation, name }),
   });
 }
 
@@ -315,6 +317,9 @@ function getClientDefaultApiVersion<
   context: SdkContext<TOptions, TServiceOperation>,
   client: SdkClient | SdkOperationGroup
 ): string | undefined {
+  if (context.apiVersion && !["latest", "all"].includes(context.apiVersion)) {
+    return context.apiVersion;
+  }
   let defaultVersion = getDefaultApiVersion(context, client.service)?.value;
   if (!defaultVersion) {
     // eslint-disable-next-line deprecation/deprecation
@@ -409,15 +414,17 @@ function getSdkMethods<TOptions extends object, TServiceOperation extends SdkSer
   for (const operationGroup of listOperationGroups(context, client)) {
     // We create a client accessor for each operation group
     const operationGroupClient = diagnostics.pipe(createSdkClientType(context, operationGroup));
+    const name = `get${operationGroup.type.name}`;
     retval.push({
       kind: "clientaccessor",
       parameters: [],
-      name: `get${operationGroup.type.name}`,
+      name,
       description: getDocHelper(context, operationGroup.type).description,
       details: getDocHelper(context, operationGroup.type).details,
       access: "internal",
       response: operationGroupClient,
       apiVersions: getAvailableApiVersions(context, operationGroup.type),
+      crossLanguageDefintionId: getCrossLanguageDefinitionId({ ...operationGroup.type, name }),
     });
   }
   return diagnostics.wrap(retval);