+++ packages/typespec-client-generator-core/CHANGELOG.md
@@ -1,5 +1,35 @@
 # Change Log - @azure-tools/typespec-client-generator-core
 
+## 0.41.8
+
+### Bug Fixes
+
+- [#753](https://github.com/Azure/typespec-azure/pull/753) fix usage propagation from sub types
+
+
+## 0.41.7
+
+### Bug Fixes
+
+- [#748](https://github.com/Azure/typespec-azure/pull/748) add crossLanguageDefinitionId onto method types
+
+
+## 0.41.6
+
+### Bug Fixes
+
+- [#741](https://github.com/Azure/typespec-azure/pull/741) use correct default api version when projecting to a specific version
+
+
+## 0.41.5
+
+### Bug Fixes
+
+- [#727](https://github.com/Azure/typespec-azure/pull/727) export int and float type judgement function
+- [#731](https://github.com/Azure/typespec-azure/pull/731) fix `@clientName` lost after adding versioning support
+- [#726](https://github.com/Azure/typespec-azure/pull/726) fix additional property union naming problem
+
+
 ## 0.41.4
 
 ### Bug Fixes